+++ appData/src/gb/src/core/vm_ui.c
@@ -163,7 +163,7 @@ void vm_overlay_clear(SCRIPT_CTX * THIS, UBYTE x, UBYTE y, UBYTE w, UBYTE h, UBY
 #ifdef CGB
         if (_is_CGB) {
             VBK_REG = 1;
-            fill_win_rect(x, y, w, h, (UI_PALETTE & 0x07u));        
+            fill_win_rect(x, y, w, h, overlay_priority | (UI_PALETTE & 0x07u));        
             VBK_REG = 0;
         }
 #endif    
@@ -185,16 +185,6 @@ void vm_choice(SCRIPT_CTX * THIS, INT16 idx, UBYTE options, UBYTE count) OLDCALL
     THIS->PC += sizeof(menu_item_t) * count;
 }
 
-void vm_load_frame(SCRIPT_CTX * THIS, UBYTE bank, UBYTE * offset) OLDCALL BANKED {
-    THIS;
-    ui_load_frame_tiles(offset, bank);
-}
-
-void vm_load_cursor(SCRIPT_CTX * THIS, UBYTE bank, UBYTE * offset) OLDCALL BANKED {
-    THIS;
-    ui_load_cursor_tile(offset, bank);
-}
-
 void vm_set_font(SCRIPT_CTX * THIS, UBYTE font_index) OLDCALL BANKED {
     THIS;
     vwf_current_font_bank = ui_fonts[font_index].bank;
@@ -213,7 +203,7 @@ void vm_overlay_scroll(SCRIPT_CTX * THIS, UBYTE x, UBYTE y, UBYTE w, UBYTE h, UB
 #ifdef CGB
     if (_is_CGB) {
         VBK_REG = 1;
-        scroll_rect(base_addr, w, h, (UI_PALETTE & 0x07u));
+        scroll_rect(base_addr, w, h, overlay_priority | (UI_PALETTE & 0x07u));
         VBK_REG = 0;
     }
 #endif
@@ -221,9 +211,10 @@ void vm_overlay_scroll(SCRIPT_CTX * THIS, UBYTE x, UBYTE y, UBYTE w, UBYTE h, UB
 
 void set_xy_win_submap(const UBYTE * source, UBYTE bank, UBYTE width, UBYTE x, UBYTE y, UBYTE w, UBYTE h) OLDCALL;
 
-void vm_overlay_set_submap(SCRIPT_CTX * THIS, UBYTE x, UBYTE y, UBYTE w, UBYTE h, UBYTE scene_x, UBYTE scene_y) OLDCALL BANKED {
-    THIS;
+void vm_overlay_set_submap(SCRIPT_CTX * THIS, INT16 x_idx, INT16 y_idx, UBYTE w, UBYTE h, UBYTE scene_x, UBYTE scene_y) OLDCALL BANKED {
     UWORD offset = (scene_y * image_tile_width) + scene_x;
+    UBYTE x = *((x_idx < 0) ? THIS->stack_ptr + x_idx : script_memory + x_idx);
+    UBYTE y = *((y_idx < 0) ? THIS->stack_ptr + y_idx : script_memory + y_idx);
 #ifdef CGB
     if (_is_CGB) {
         VBK_REG = 1;
@@ -234,8 +225,10 @@ void vm_overlay_set_submap(SCRIPT_CTX * THIS, UBYTE x, UBYTE y, UBYTE w, UBYTE h
     set_xy_win_submap(image_ptr + offset, image_bank, image_tile_width, x, y, w, h);
 }
 
-void vm_overlay_set_map(SCRIPT_CTX * THIS, INT16 idx, UBYTE x, UBYTE y, UBYTE bank, const background_t * background) OLDCALL BANKED {
+void vm_overlay_set_map(SCRIPT_CTX * THIS, INT16 idx, INT16 x_idx, INT16 y_idx, UBYTE bank, const background_t * background) OLDCALL BANKED {
     far_ptr_t tilemap;
+    UBYTE x = *((x_idx < 0) ? THIS->stack_ptr + x_idx : script_memory + x_idx);
+    UBYTE y = *((y_idx < 0) ? THIS->stack_ptr + y_idx : script_memory + y_idx);
     UBYTE w = ReadBankedUBYTE((void *)&(background->width), bank);
     UBYTE h = ReadBankedUBYTE((void *)&(background->height), bank);
     _map_tile_offset = *(INT16 *)(VM_REF_TO_PTR(idx));
@@ -254,10 +247,9 @@ void vm_overlay_set_map(SCRIPT_CTX * THIS, INT16 idx, UBYTE x, UBYTE y, UBYTE ba
     _map_tile_offset = 0;
 }
 
-void vm_set_text_sound(SCRIPT_CTX * THIS, UBYTE frames, UBYTE channel) OLDCALL BANKED {
-    text_sound_frames = frames;
-    text_sound_ch = channel; 
-    text_sound_bank = THIS->bank;  
-    text_sound_data = THIS->PC;
-    THIS->PC += ((channel == 3) ? 0x15 : 5); // skip regs and waveform, if playing on ch3
+void vm_set_text_sound(SCRIPT_CTX * THIS, UBYTE bank, UBYTE * offset, UBYTE channel_mask) OLDCALL BANKED {
+    THIS;
+    text_sound_bank = bank;  
+    text_sound_data = offset;
+    text_sound_mask = channel_mask; 
 }